# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VentasFix",
        "description": "Microservicio de carro de compra de VentasFix: usuarios, productos, clientes y dashboard. Autenticación con JWT (Bearer).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "3.1 Listar todos los clientes",
                "operationId": "413c12320d4121431e38339f45ecaa7c",
                "responses": {
                    "200": {
                        "description": "Listado obtenido"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "3.3 Agregar un nuevo cliente",
                "operationId": "f909473745c1a1d469e0c5a5b54682b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_rut",
                                    "business_sector",
                                    "business_name",
                                    "phone",
                                    "address",
                                    "contact_name",
                                    "contact_email"
                                ],
                                "properties": {
                                    "company_rut": {
                                        "type": "string",
                                        "example": "78.111.222-3"
                                    },
                                    "business_sector": {
                                        "type": "string",
                                        "example": "Minería"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "Minera del Norte S.A."
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+56 9 8765 4321"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Ruta 5 Norte km 12, Antofagasta"
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "example": "Carla Muñoz"
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "example": "cmunoz@mineranorte.cl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado"
                    },
                    "422": {
                        "description": "Datos inválidos o vacíos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "3.2 Obtener un cliente por su ID",
                "operationId": "feff36e8c89b21277ad216c29ca9ca4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "No existe un cliente con ese id"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "3.4 Actualizar un cliente por su ID (PUT o PATCH)",
                "operationId": "4cdaf0a748c704e6f11a98f7bb6eb5a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company_rut": {
                                        "type": "string",
                                        "example": "76.123.456-7"
                                    },
                                    "business_sector": {
                                        "type": "string",
                                        "example": "Retail y distribución"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "Comercial Sur SpA"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+56 9 1234 5678"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Libertador 1234, Santiago"
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "example": "Luis Soto"
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "example": "luis.soto@comercialsur.cl"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado"
                    },
                    "404": {
                        "description": "No existe un cliente con ese id"
                    },
                    "422": {
                        "description": "Datos inválidos o vacíos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "3.5 Eliminar un cliente por su ID",
                "operationId": "5b6d3bdaf49885db3b7b4b23aed2156b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado"
                    },
                    "404": {
                        "description": "No existe un cliente con ese id"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Pega el access_token q devuelve /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}